refactor: moves WalletContent
